Move `loadMetadataForCard` to questions module
