+++ enterprise/frontend/src/metabase-enterprise/audit_app/pages/AuditQueryDetail.jsx
@@ -54,7 +54,7 @@ const AuditQueryDetail = ({ params: { queryHash } }) => (
 import { getMetadata } from "metabase/selectors/metadata";
 
 import ExplicitSize from "metabase/components/ExplicitSize";
-import { loadMetadataForCard } from "metabase/query_builder/actions";
+import { loadMetadataForCard } from "metabase/questions/actions";
 import NativeQuery from "metabase-lib/queries/NativeQuery";
 import Question from "metabase-lib/Question";
 import * as QueryDetailCards from "../lib/cards/query_detail";
