+++ frontend/src/metabase/query_builder/actions/core/index.ts
@@ -1,4 +1,3 @@
 export * from "./core";
 export * from "./initializeQB";
-export * from "./metadata";
 export * from "./updateQuestion";
